move useAssertion hook to utils
